Attack Code (Code Finished, See notes about Animation)
Attack Code is done. All that's left is to code in the damage to other entities, which I'll do when we get some enemies on the scene.
ANIMATION NOTE: The attack is currently invisible. I'm going to hook up the attack stance to the player, but could we get a sort of trail for the swing?